+++ src/sac/Objet.java
@@ -5,12 +5,12 @@ public class Objet
 	private String _nom;
 	public String nom() { return this._nom; }
 	
-	private float _poids;
-	public float poids() { return this._poids; }
-	private float _valeur;
-	public float valeur() { return this._valeur; }
+	private double _poids;
+	public double poids() { return this._poids; }
+	private double _valeur;
+	public double valeur() { return this._valeur; }
 
-	public Objet(String nom, float poids, float valeur)
+	public Objet(String nom, double poids, double valeur)
 	{
 		this._nom = nom;
 		this._poids = poids;
